Renamed the mixin module
