+++ lib/msf/core/reflective_dll_injection.rb
