+++ modules/exploits/windows/local/ms10_015_kitrap0d.rb
@@ -4,7 +4,7 @@
 ##
 
 require 'msf/core'
-require 'msf/core/rdi_mixin'
+require 'msf/core/reflective_dll_injection'
 require 'msf/core/exploit/exe'
 require 'rex'
