+++ modules/exploits/windows/local/ppr_flatten_rec.rb
@@ -4,7 +4,7 @@
 ##
 
 require 'msf/core'
-require 'msf/core/rdi_mixin'
+require 'msf/core/reflective_dll_injection'
 require 'rex'
 
 class Metasploit3 < Msf::Exploit::Local
